+++ tests/test_generic_search_cython.py
@@ -1,4 +1,6 @@
-from tests.compat import b, unittest
+import unittest
+
+from tests.compat import b
 from tests.utils import skip_if_arguments_arent_byteslike
 from tests.test_levenshtein import TestFindNearMatchesLevenshteinBase
 from fuzzysearch.common import Match, get_best_match_in_group, group_matches, LevenshteinSearchParams