+++ tests/utils.py
@@ -1,15 +1,13 @@
 from functools import wraps
 
-from fuzzysearch.compat import text_type
-
 
 def skip_if_arguments_arent_byteslike(test_method):
     @wraps(test_method)
     def new_method(self, *args, **kwargs):
         subsequence, sequence = args[:2]
         if (
-            isinstance(subsequence, text_type) or
-            isinstance(sequence, text_type)
+            isinstance(subsequence, str) or
+            isinstance(sequence, str)
         ):
             raise self.skipTest(
                 "skipping test with unicode data for byteslike function")